optimize case of no overlapping ranges
